modles/requirements.txt
deep_sort_realtime==1.3.2
matplotlib==3.8.3
numpy==1.26.4
onnx==1.16.0
onnxsim==0.4.36
opencv_python==4.9.0.80
opencv_python_headless==4.9.0.80
Pillow==10.3.0
pycocotools==2.0.7
scipy==1.13.0
torch==2.2.1
torchvision==0.17.1
tqdm==4.66.2
visdom==0.2.4
